Update to latest lufa version

+++ avr/bootloaders/HoodLoader2/Descriptors.h
@@ -175,9 +175,8 @@ along with Hoodloader2.  If not, see <http://www.gnu.org/licenses/>.
 
 	/* Function Prototypes: */
 		uint16_t CALLBACK_USB_GetDescriptor(const uint16_t wValue,
-		                                    const uint8_t wIndex,
+		                                    const uint16_t wIndex,
 		                                    const void** const DescriptorAddress)
 		                                    ATTR_WARN_UNUSED_RESULT ATTR_NON_NULL_PTR_ARG(3);
 
 #endif
-
